+++ packages/arui-scripts/src/configs/babel-dependencies.ts
@@ -1,0 +1,34 @@
+import applyOverrides from './util/apply-overrides';
+import browsers from './supporting-browsers';
+
+export const babelDependencies = applyOverrides('babelDependencies', {
+    sourceType: 'unambiguous',
+    presets: [
+        [
+            require.resolve('@babel/preset-env'),
+            {
+                // Allow importing core-js in entrypoint and use browserlist to select polyfills
+                useBuiltIns: 'entry',
+                // Set the corejs version we are using to avoid warnings in console
+                // This will need to change once we upgrade to corejs@3
+                corejs: 3,
+                // Exclude transforms that make all code slower
+                exclude: ['transform-typeof-symbol'],
+            },
+        ]
+    ],
+    plugins: [
+        [
+            require.resolve('@babel/plugin-transform-runtime'),
+            {
+                corejs: false,
+                helpers: true,
+                // By default, babel assumes babel/runtime version 7.0.0-beta.0,
+                // explicitly resolving to match the provided helper functions.
+                // https://github.com/babel/babel/issues/10261
+                version: require('@babel/runtime/package.json').version,
+                regenerator: true,
+            },
+        ]
+    ]
+});